+++ public/data/feed.json
@@ -8,7 +8,14 @@
       "heartImg": "https://images.unsplash.com/photo-1546527868-ccb7ee7dfa6a?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1170&q=80",
       "heartUser": "youworld",
       "heartUsers": "10",
-      "feedContent": "너-무 행복했던 4월 1일💕"
+      "feedContent": "너-무 행복했던 4월 1일💕",
+      "comment": [
+        {
+          "id": 1,
+          "commentUser": "qxxxqwwi",
+          "commentContent": "👍👍"
+        }
+      ]
     },
     {
       "id": 2,
@@ -18,7 +25,19 @@
       "heartImg": "https://images.unsplash.com/photo-1487412720507-e7ab37603c6f?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1171&q=80",
       "heartUser": "candy_lp",
       "heartUsers": "7",
-      "feedContent": "👍"
+      "feedContent": "👍",
+      "comment": [
+        {
+          "id": 1,
+          "commentUser": "jelly",
+          "commentContent": "해삐해삐"
+        },
+        {
+          "id": 2,
+          "commentUser": "icecream",
+          "commentContent": "✨✨✨"
+        }
+      ]
     },
     {
       "id": 3,
@@ -28,7 +47,24 @@
       "heartImg": "https://images.unsplash.com/photo-1526336024174-e58f5cdd8e13?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxzZWFyY2h8OHx8JUVBJUIzJUEwJUVDJTk2JTkxJUVDJTlEJUI0fGVufDB8fDB8fA%3D%3D&auto=format&fit=crop&w=500&q=60",
       "heartUser": "sunny",
       "heartUsers": "8",
-      "feedContent": "바다🐳"
+      "feedContent": "바다🐳",
+      "comment": [
+        {
+          "id": 1,
+          "commentUser": "qxxxqwwi",
+          "commentContent": "바다다~"
+        },
+        {
+          "id": 2,
+          "commentUser": "sunny",
+          "commentContent": "🐳🐳🐳"
+        },
+        {
+          "id": 3,
+          "commentUser": "icecandy",
+          "commentContent": "짱 예뻐"
+        }
+      ]
     },
     {
       "id": 4,
@@ -38,7 +74,19 @@
       "heartImg": "https://images.unsplash.com/photo-1618359057154-e21ae64350b6?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=2070&q=80",
       "heartUser": "happiness",
       "heartUsers": "12",
-      "feedContent": "아이스크림🍦"
+      "feedContent": "아이스크림🍦",
+      "comment": [
+        {
+          "id": 1,
+          "commentUser": "qxxxqwwi",
+          "commentContent": "나도"
+        },
+        {
+          "id": 2,
+          "commentUser": "ufowlao",
+          "commentContent": "우왕"
+        }
+      ]
     },
     {
       "id": 5,
@@ -48,7 +96,14 @@
       "heartImg": "https://images.unsplash.com/photo-1439405326854-014607f694d7?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxzZWFyY2h8NXx8JUVCJUIwJTk0JUVCJThCJUE0fGVufDB8fDB8fA%3D%3D&auto=format&fit=crop&w=500&q=60",
       "heartUser": "nice_lk",
       "heartUsers": "8",
-      "feedContent": "🎆✨"
+      "feedContent": "🎆✨",
+      "comment": [
+        {
+          "id": 1,
+          "commentUser": "qxxxqwwi",
+          "commentContent": "예쁘다"
+        }
+      ]
     }
   ]
 }
